pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docscope"
version = "1.0.0"
description = "Universal Documentation Browser & Search System"
readme = "README.md"
authors = [
    {name = "DocScope Team", email = "team@docscope.io"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Documentation",
    "Topic :: Text Processing :: Indexing",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0",
    "whoosh>=2.7",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0",
    "httpx>=0.24.0",
    "python-dateutil>=2.8",
    "watchdog>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "httpx>=0.24.0",
]
ai = [
    "openai>=1.0",
    "tiktoken>=0.5",
    "numpy>=1.24",
    "scikit-learn>=1.3",
]
ocr = [
    "pytesseract>=0.3",
    "pdf2image>=1.16",
    "pillow>=10.0",
]
enterprise = [
    "python-jose[cryptography]>=3.3",
    "python-ldap>=3.4",
    "redis>=5.0",
    "prometheus-client>=0.18",
]
all = [
    "docscope[dev,ai,ocr,enterprise]",
]

[project.scripts]
docscope = "docscope.cli:main"

[project.urls]
Homepage = "https://github.com/docscope/docscope"
Documentation = "https://docs.docscope.io"
Repository = "https://github.com/docscope/docscope"
Issues = "https://github.com/docscope/docscope/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["docscope*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "B", "UP"]
ignore = ["E501"]
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=docscope --cov-report=term-missing"